### 2023-05-21 #5486 @editorialbot
[PRE REVIEW]: MLMOD: Machine Learning Methods for Data-Driven Modeling in LAMMPS
**Submitting author:** @atzberg (<a href="http://orcid.org/0000-0001-6806-8069">Paul J. Atzberger</a>)
**Repository:** https://github.com/atzberg/mlmod
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
