+++ paper/paper.bib
@@ -141,6 +141,7 @@
   publisher =     {Wiley Online Library},
   title =         {CHARMM: a program for macromolecular energy,
                    minimization, and dynamics calculations},
+  doi     = {10.1002/jcc.540040211},
   volume =        {4},
   year =          {1983},
 }
@@ -156,6 +157,7 @@
   pages     = {1--19},
   volume    = {117},
   timestamp = {2015.06.12},
+  doi       = {10.2172/10176421},
   url       = {http://www.sciencedirect.com/science/article/pii/
                   S002199918571039X},
 }
@@ -177,6 +179,7 @@
   number  = {7567},
   pages   = {47--55},
   volume  = {525},
+  doi     = {10.1038/nature14956},
   url     = {https://doi.org/10.1038/nature14956},
 }
 
@@ -190,6 +193,7 @@
   number  = {3},
   pages   = {169--174},
   volume  = {36},
+  doi     = {10.1557/mrs.2011.30},
   url     = {https://doi.org/10.1557/mrs.2011.30},
 }
 
@@ -203,6 +207,7 @@
                    atom simulations of nanoscale systems},
   volume =        {157},
   year =          {2007},
+  doi  = {10.1016/j.jsb.2006.10.023},
 }
 
 @Article{Washington:2009,
@@ -228,6 +233,7 @@
   volume =        {1},
   year =          {2021},
   issn =          {2662-8457},
+  doi  =          {10.1038/s43588-021-00034-x},
   url =           {https://doi.org/10.1038/s43588-021-00034-x},
 }
 
@@ -349,6 +355,7 @@
   volume =        {633},
   year =          {1941},
   timestamp =     {2017.04.17},
+  doi     = {10.1016/0079-6816(93)90013-l},
 }
 
 @book{Doi:2013,
