+++ paper/paper.bib
@@ -69,7 +69,7 @@
   issn    = {0167-2789},
   pages   = {57-70},
   volume  = {265},
-  doi     = {https://doi.org/10.1016/j.physd.2013.09.002},
+  doi     = {10.1016/j.physd.2013.09.002},
   url     = {https://www.sciencedirect.com/science/article/pii/
                   S016727891300256X},
 }
