+++ paper/paper.bib
@@ -70,8 +70,7 @@
   pages   = {57-70},
   volume  = {265},
   doi     = {10.1016/j.physd.2013.09.002},
-  url     = {https://www.sciencedirect.com/science/article/pii/
-                  S016727891300256X},
+  url     = {https://doi.org/10.1016/j.physd.2013.09.002},
 }
 
 @book{Bird:1987,
