### 2023-05-22 #5486 @atzberg
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1126/science.1127647 is OK
> - 10.1073/pnas.1517384113 is OK
> - 10.1126/science.1165893 is OK
> - 10.1098/rsta.2008.0219 is OK
> - 10.1007/978-3-319-01905-5_60-2 is OK
> - 10.1088/1361-651x/ab7150 is OK
> - 10.1007/978-3-540-28650-9_4 is OK
> - 10.1021/acsomega.8b01393 is OK
> - 10.5281/zenodo.7897824 is OK
> - 10.1080/01621459.2017.1285773 is OK
>
> MISSING DOIs
>
> - 10.1002/jcc.540040211 may be a valid DOI for title: CHARMM: a program for macromolecular energy, minimization, and dynamics calculations
> - 10.2172/10176421 may be a valid DOI for title: Fast Parallel Algorithms for Short-Range Molecular Dynamics
> - 10.1038/nature14956 may be a valid DOI for title: The quiet revolution of numerical weather prediction
> - 10.1557/mrs.2011.30 may be a valid DOI for title: High-performance computing for materials design to advance energy science
> - 10.1016/j.jsb.2006.10.023 may be a valid DOI for title: High performance computing in biology: multimillion atom simulations of nanoscale systems
> - 10.1038/s43588-021-00034-x may be a valid DOI for title: Scaling up system size in materials simulation
> - 10.1016/0079-6816(93)90013-l may be a valid DOI for title: Theory of the stability of strongly charged lyophobic sols and of the adhesion of strongly charged particles in solutions of electrolytes
>
> INVALID DOIs
>
> - https://doi.org/10.1016/j.physd.2013.09.002 is INVALID because of 'https://doi.org/' prefix
> ```

The DOIs have now been fixed.
