### 2023-06-23 #5486 @ppxasjsm
@atzberg the first task I have for you is to look at the missing and invalid DOIs in your paper. Can you try and fix these please.
