### 2023-06-23 #5486 @atzberg
@ppxasjsm <https://github.com/ppxasjsm> Thanks!  Concerning DOI's I saw the
bot's comments on missing doi's.  I made edits a few weeks ago to resolve
links and doi's to all of the referenced papers.  Some of the papers did
not have DOIs, so I tried to include web link to the archived paper.  Just
let me know if the current edits look alright or if something else is
needed.
