### 2023-06-27 #5486 @atzberg
@ppxasjsm <https://github.com/ppxasjsm>  Thanks again!  The .md compiled
fine with the docker image provided by JOSS.  I put a PDF of the paper in
the folder /paper along with the markdown.  Just let me know if you need
any other information or items.
