+++ paper/paper.bib
@@ -1,4 +1,3 @@
-% Encoding: UTF-8
 @article{Hinton:2006,
   author =        {Geoffrey Hinton and Ruslan Salakhutdinov},
   journal =       {Science},
@@ -73,14 +72,13 @@
   url     = {https://doi.org/10.1016/j.physd.2013.09.002},
 }
 
-@book{Bird:1987,
-  author =          {Bird,R.B., Curtiss, C. F., Armstrong, R. C., Hassager,
-  O.},
-  publisher =     {Wiley-Interscience},
-  title =         {Dynamics of Polymeric Liquids : Volume II Kinetic
-                   Theory},
-  year =          {1987},
-  timestamp =     {2009.07.29},
+@Book{Bird:1987,
+  author    = {Bird,R.B., Curtiss, C. F., Armstrong, R. C., Hassager, O.},
+  publisher = {Wiley-Interscience},
+  title     = {Dynamics of Polymeric Liquids : Volume II Kinetic Theory},
+  year      = {1987},
+  timestamp = {2009.07.29},
+  url       = {https://www.wiley.com/en-us/Dynamics+of+Polymeric+Liquids,+Volume+2:+Kinetic+Theory,+2nd+Edition-p-9780471802440},
 }
 
 @Article{Lubensky:1997,
@@ -92,7 +90,9 @@
   number    = {2-3},
   pages     = {187-197--},
   volume    = {102},
+  doi       = {10.1016/S0038-1098(96)00718-1},
   timestamp = {2009.07.14},
+  url       = {https://doi.org/10.1016/S0038-1098(96)00718-1},
 }
 
 @Article{Kimura:2009,
@@ -105,16 +105,19 @@
   pages     = {8--},
   volume    = {78},
   address   = {Kyushu Univ, Dept Phys, Sch Sci, Higashi Ku, Fukuoka 8128581, Japan. Kimura, Y, Kyushu Univ, Dept Phys, Sch Sci, Higashi Ku, 6-10-1 Hakozaki, Fukuoka 8128581, Japan. kimura@phys.kyushu-u.ac.jp},
+  doi       = {10.1143/JPSJ.78.041005},
   timestamp = {2009.07.14},
+  url       = {https://doi.org/10.1143/JPSJ.78.041005},
 }
 
-@book{Mccammon:1988,
-  author =        {McCammon, J.A. and Harvey, S.C.},
-  publisher =     {Cambridge University Press},
-  title =         {Dynamics of Proteins and Nucleic Acids},
-  year =          {1988},
-  isbn =          {9780521356527},
-  url =           {https://books.google.com/books?id=OXDS3Oq2QYIC},
+@Book{Mccammon:1988,
+  author    = {McCammon, J.A. and Harvey, S.C.},
+  publisher = {Cambridge University Press},
+  title     = {Dynamics of Proteins and Nucleic Acids},
+  year      = {1988},
+  isbn      = {9780521356527},
+  doi       = {10.1017/CBO9781139167864},
+  url       = {https://doi.org/10.1017/CBO9781139167864},
 }
 
 @Article{Karplus:2002,
@@ -161,11 +164,12 @@
                   S002199918571039X},
 }
 
-@book{Richardson:2007,
-  author =        {Richardson, Lewis Fry},
-  publisher =     {Cambridge university press},
-  title =         {Weather prediction by numerical process},
-  year =          {2007},
+@Book{Richardson:2007,
+  author    = {Richardson, Lewis Fry},
+  publisher = {Cambridge university press},
+  title     = {Weather prediction by numerical process},
+  year      = {2007},
+  url       = {https://archive.org/details/weatherpredictio00richrich},
 }
 
 @Article{Bauer:2015,
@@ -271,6 +275,7 @@
   address   = {Cambridge, MA, USA},
[…]
