### 2023-06-27 #5486 @atzberg
For the papers not having existing DOIs, I have added related URLs.
