+++ paper/paper.bib
@@ -397,17 +397,14 @@
   url       = {https://doi.org/10.1021/acsomega.8b01393},
 }
 
-@article{Smoluchowski:1906,
-  author =        {Smoluchowski, VMV},
-  journal =       {Ann. Phys},
-  pages =         {756},
-  title =         {Drei Vorträge über Diffusion, Brownsche
-                   Molekularbewegung und Koagulation von
-                   Kolloidteilchen},
-  volume =        {21},
-  year =          {1906},
-  url =           {https://jbc.bj.uj.edu.pl/Content/387533/PDF/
-                  FIZART_SMOLUCHOWSKI_00093.pdf},
+@Article{Smoluchowski:1906,
+  author  = {Smoluchowski, VMV},
+  journal = {Ann. Phys},
+  title   = {Drei Vorträge über Diffusion, Brownsche Molekularbewegung und Koagulation von Kolloidteilchen},
+  year    = {1906},
+  pages   = {756},
+  volume  = {21},
+  url     = {https://jbc.bj.uj.edu.pl/Content/387533/PDF/FIZART_SMOLUCHOWSKI_00093.pdf},
 }
 
 @article{Atzberger:2020,
