### 2023-06-27 #5486 @atzberg
Thanks, unfortunately the editorialbot suggestion "MISSING DOIs -
10.2139/ssrn.4455789 may be a valid DOI for title: SDYN-GANs: Adversarial
Learning Methods for Multistep Generative Models for General Order
Stochastic Dynamics"  is not correct, since this appears to be a website
aggregator that grabbed our preprint from someplace on-line.  The correct
reference is the arXiv link we have now added as URL for the preprint
manuscript.   The references should now be done, with each also having
either a URL or a DOI when those exist and possible to include.  Just let
me know if you need any other edits or information.  -- Paul
