### 2023-07-04 #5486 @ppxasjsm
👋 @RMeli and  @dirmeier would you be willing to review this submission for JOSS?
