### 2023-07-05 #5620 @editorialbot
[REVIEW]: MLMOD: Machine Learning Methods for Data-Driven Modeling in LAMMPS
**Submitting author:** @atzberg (<a href="http://orcid.org/0000-0001-6806-8069">Paul J. Atzberger</a>)
**Repository:** https://github.com/atzberg/mlmod
**Branch with paper.md** (empty if default branch): main
**Version:** 1.0.1
