### 2023-08-04 #5620 @JColdstream
@atzberg
The example runs and gives me a nice trajectory, but have you got any checks to make sure the code is actually giving the right answer?
Either something similar to pytest or something like an example calculation/dataset to compare to maybe?
Just looking for something for the Automated tests section in the review.

Other things:
I really like the little intro video on YouTube.
LAMMPS has a new citation that they prefer to be used (https://doi.org/10.1016/j.cpc.2021.108171) rather than the old Plimpton 1995 one.
