### 2023-08-11 #5620 @RuiApostolo
I'll try to organise my recommendations in relevant sections

# Installation

## pip

The installation using pip is really simple, but it contains an older version of LAMMPS (29 Oct 2020).
If possible, it would be preferable to upload the wheel file to pypi.org so that users could simply `pip install mlmod`.

# Compiling

Compiling MLMOD from source has quite a few steps, and while it's not overly complicated, I would not recommend it for anyone not used to compiling software with (mildly) complex dependencies.
It could be made easier if there was a more detailed example script containing all the commands and some example paths.
In Ubuntu 22.04, FFTW3 (libfftw3-dev), LIBLAPACKE (liblapacke-dev), and tinyxml2 (libtinyxml2-dev) are available to install as packages, leaving the user to install only libtorch (which is very simple) and the MLMOD-patched version of LAMMPS.

With the recent change in LAMMPS development choice to provide a [newbie-friendly `cmake` compilation instructions rather than make](https://docs.lammps.org/Build_cmake.html#advantages-of-using-cmake), it would be useful to provide instructions for compiling MLMOD using `cmake`.
Furthermore, having a shell script that copies all the files MLMOD needs (src/USER-MLMOD, lib, MAKE) to a user-specified location where the LAMMPS source files are would remove several steps and sources of potential error.
Finally, consider submitting the USER-MLMOD package to the [LAMMPS repository](https://github.com/lammps/lammps/blob/6422565048106045443f022b26bd144d23505533/.github/CONTRIBUTING.md), so that it may be included in the general release of LAMMPS.

## README

On the main README, you should link the sentence "For more information on other ways to install or compile the package, please see the documentation page." to the `src/INSTALL.md` file.

# Tests

I agree with @JColdstream here, some tests to make sure that the package is returning expected results would be very useful.
It could help identify compatibility issues with certain LAMMPS or library versions, should they exist.

# Software documentation

I think this needs to be created. At a minimum, there needs to be documentation on the LAMMPS commands that MLMOD introduces, and all possible parameters they accept. Ideally, these would follow the same structure that the LAMMPS documentation already has, so that the familiar structure would help users  already familiar with LAMMPS.

# Paper

Overall well written, I have only found a few minor typos (I'm providing the lines as numbered on the PDF created by the bot above):
- `accomodated` line 32
- `accomodate` line 59
- `simiplified` line 93
- `Nielsen:2000` line 110 (missing `@` to make it a reference, I believe)
