+++ lib/mlmod/driver_mlmod.h
@@ -1,386 +1,0 @@
-/* -----------------------------------------------------------------------------
-
-  MLMOD: Machine Learning (ML) Modeling (MOD) and Simulation Package 
-
-  Paul J. Atzberger
-  http://atzberger.org/
-  
-  Please cite the follow paper when referencing this package
-  
-  "MLMOD Package: Machine Learning Methods for Data-Driven Modeling in LAMMPS",
-  Atzberger, P. J., arXiv:2107.14362, 2021.
-  
-  @article{mlmod_atzberger,
-    author  = {Atzberger, P. J.},
-    journal = {arxiv},
-    title   = {MLMOD Package: Machine Learning Methods for Data-Driven Modeling in LAMMPS},
-    year    = {2021},
-    note    = {http://atzberger.org},
-    doi     = {10.48550/arXiv.2107.14362},
-    url     = {https://arxiv.org/abs/2107.14362},
-  }
-    
-  For latest releases, examples, and additional information see 
-  http://atzberger.org/
- 
---------------------------------------------------------------------------------
-*/
-
-#ifndef LMP_DRIVER_MLMOD_H
-#define LMP_DRIVER_MLMOD_H
-
-#ifndef FFT_FFTW /* determine if FFTW package specified */
-  #warning "No FFTW package specified for MLMOD codes.  Some MLMOD functionality may be disabled."
-#else
-  #define USE_PACKAGE_FFTW3
-#endif
-
-#include "fix_mlmod.h"
-
-// PyTorch C/C++ interface
-#include <torch/torch.h>
-
-#include "tinyxml2.h"
-
-//#include "mpi.h"
-#include "fix.h"
-#include "lammps.h"
-#include "random_mars.h"
-
-#include <cstdlib>
-#include <cstdio>
-#include <cstring>
-#include <cstddef>
-
-using namespace std;
-using namespace tinyxml2;
-using namespace LAMMPS_NS;
-
-namespace USER_MLMOD {
-
-struct DriverMLMOD {
-
- public:
-
-  /* ================= function prototypes ================= */
-  DriverMLMOD(class FixMLMOD *, class LAMMPS *, int, char **);
-
-  DriverMLMOD(); /* used to construct empty object,
-                primarily for testing purposes */
-  ~DriverMLMOD();
-
-  // =========================== Fix Related Function Calls ==================== */
-  int          setmask();
-  virtual void init();
-  void         setup(int vflag);
-  virtual void initial_integrate(int);
-  virtual void final_integrate();  
-  void         reset_dt();
-  void         post_force(int vflag);
-  
-  // additional methods/hooks
-  void pre_exchange();
-  void end_of_step();
-
-  /* =========================== MLMOD Function Calls =========================== */
-  void init_attributes();
-  void init_from_fix();
-  void constr_input_masked(at::TensorAccessor<float,2,at::DefaultPtrTraits,long> input_m_a,
-                           int num_indices,int *indices,
-                           int mask_input, int mask_list_n, int *mask_list,
-                           double **x,double **v,double **f, int *type);
-  void parse_xml_params(char *filename);
-  
-  /* write the MLMOD simulation data to disk */
-  //void writeAllSimulationData(int timeIndex);
-  void writeInfo();
-  void writeFinalInfo();
-
-  /* case specific functions */ 
-  void initial_integrate_dX_MF_Q1_ML1_Pair();  
-  void initial_integrate_dX_MF_Q1_ML1_N2N();  
-  void initial_integrate_F_ML1();  
[…]
