+++ lib/mlmod/build.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-
-echo "Building mlmod library"
-
-make -f Makefile.serial clean
-make -f Makefile.serial lib_shared
-
-
