+++ src/lib/mlmod/driver_mlmod.h
@@ -1,0 +1,504 @@
+/**
+ * @file 
+ *
+ * @brief MLMOD library routines to drive the machine learning models 
+ * and simulations.
+ *
+ * 
+ */
+
+/* -----------------------------------------------------------------------------
+
+  MLMOD: Machine Learning (ML) Modeling (MOD) and Simulation Package 
+
+  Paul J. Atzberger
+  http://atzberger.org/
+  
+  Please cite the follow paper when referencing this package
+  
+  "MLMOD Package: Machine Learning Methods for Data-Driven Modeling in LAMMPS",
+  Atzberger, P. J., arXiv:2107.14362, 2021.
+  
+  @article{mlmod_atzberger,
+    author  = {Atzberger, P. J.},
+    journal = {arxiv},
+    title   = {MLMOD Package: Machine Learning Methods for Data-Driven Modeling in LAMMPS},
+    year    = {2021},
+    note    = {http://atzberger.org},
+    doi     = {10.48550/arXiv.2107.14362},
+    url     = {https://arxiv.org/abs/2107.14362},
+  }
+    
+  For latest releases, examples, and additional information see 
+  http://atzberger.org/
+ 
+--------------------------------------------------------------------------------
+*/
+
+#ifndef LMP_DRIVER_SELM_H
+#define LMP_DRIVER_SELM_H
+
+#ifndef FFT_FFTW /* determine if FFTW package specified */
+  #warning "No FFTW package specified for SELM codes.  The SELM functionality will be disabled."
+#else
+  #define USE_PACKAGE_FFTW3
+#endif
+
+#include "fix_mlmod.h"
+
+// PyTorch C/C++ interface
+#include <torch/torch.h>
+
+#include "tinyxml2.h"
+
+//#include "mpi.h"
+#include "fix.h"
+#include "lammps.h"
+#include "random_mars.h"
+
+#include <cstdlib>
+#include <cstdio>
+#include <cstring>
+#include <cstddef>
+
+using namespace std;
+using namespace tinyxml2;
+using namespace LAMMPS_NS;
+
+namespace USER_MLMOD {
+
+/**
+ * @brief The driver of the functionality for the shared library class.  This
+ * specifies the prototype for the API interface with lammps.
+ * 
+ */
+struct DriverMLMOD {
+
+ public:
+
+  /* ================= function prototypes ================= */
+
+  /** 
+   * @brief Main constructor for mlmod package driver. 
+   *
+   * @param FixMLMOD*: reference to the mlmod-lammps "fix" interface  
+   * @param LAMMPS*: reference to the running lammps instance
+   * @param narg: number of command arguments
+   * @param args: list of strings for the arguments 
+   *
+  */
+  DriverMLMOD(class FixMLMOD *, class LAMMPS *, int, char **);
+
+  /** 
+   * @brief Constructor for empy object which is used primarily for testing. 
+  */
+  DriverMLMOD(); /* used to construct empty object,
+                    primarily for testing purposes */
+
+  /** The deconstructor. */
+  ~DriverMLMOD();
+
+  // =========================== Fix Related Function Calls ==================== */
+  /** Lammps interface code for signaling the type of simulation. */
+  int          setmask();
+
+  /** Used for initializing the driver settings when created. */
+  virtual void init();
+
+  /** Used for setting up references in the driver after lammps is initialized. */
+  void         setup(int vflag);
+
+  /** Lammps calls this routine at the start of each time-step. */
+  virtual void initial_integrate(int);
+
+  /** Lammps calls this routine at the end of each time-step. */
+  virtual void final_integrate();  
+
[…]
