+++ doc/README.md
@@ -1,0 +1,3 @@
+- [Documentation pages:](http://web.math.ucsb.edu/~atzberg/mlmod/docs/index.html) for installation, usage, and other information. 
+- [Examples folder:](https://github.com/atzberg/mlmod/tree/master/examples) for setting up models and simulations.
+
