+++ src/USER-MLMOD/fix_mlmod.cpp
@@ -1,205 +1,0 @@
-/* -----------------------------------------------------------------------------
-
-  MLMOD: Machine Learning (ML) Modeling (MOD) and Simulation Package 
-
-  Paul J. Atzberger
-  http://atzberger.org/
-  
-  Please cite the follow paper when referencing this package
- 
-  "MLMOD Package: Machine Learning Methods for Data-Driven Modeling in LAMMPS",
-  Atzberger, P. J., arXiv:2107.14362, 2021.
-  
-  @article{mlmod_atzberger,
-    author  = {Atzberger, P. J.},
-    journal = {arxiv},
-    title   = {MLMOD Package: Machine Learning Methods for Data-Driven Modeling in LAMMPS},
-    year    = {2021},
-    note    = {http://atzberger.org},
-    doi     = {10.48550/arXiv.2107.14362},
-    url     = {https://arxiv.org/abs/2107.14362},
-  }
-    
-  For latest releases, examples, and additional information see 
-  http://atzberger.org/
-
---------------------------------------------------------------------------------
-*/
-
-#include <cstdlib>
-#include <cstddef>
-#include <cstdio>
-#include <cstring>
-#include <iostream>
-#include <sstream>
-
-/* LAMMPS includes */
-#include "atom.h"
-#include "force.h"
-#include "update.h"
-#include "respa.h"
-#include "error.h"
-#include "comm.h"
-#include "universe.h"
-#include "version.h" 
-#include "random_mars.h"
-#include "citeme.h"
-
-/* MLMOD_includes */
-#include "fix_mlmod.h"
-#include "wrapper/wrapper_mlmod.h"
-
-//using namespace MLMOD;
-using namespace LAMMPS_NS;
-using namespace FixConst;
-using namespace std;
-
-static const char cite_mlmod_str[] =
-  "USER-MLMOD Package: \n"
-  "@article{mlmod_atzberger,\n"
-  "  author  = {Atzberger, P. J.},\n"
-  "  journal = {arxiv},\n"
-  "  title   = {MLMOD Package: Machine Learning Methods for " 
-  "Data-Driven Modeling in LAMMPS},\n"
-  "  year    = {2021},\n"
-  "  note    = {http://atzberger.org},\n"
-  "  doi     = {10.48550/arXiv.2107.14362},\n"
-  "  url     = {https://arxiv.org/abs/2107.14362},\n"
-  "}\n\n";
-
-/* =========================== Class definitions =========================== */
-FixMLMOD::FixMLMOD() : Fix(NULL, 0, NULL) {
-  /* WARNING: May need to modify LAMMPS codes so that we have
-              Fix(NULL, 0, NULL) acts like empty constructor */
-
-  //wrapper_mlmod = new WrapperMLMOD();
-   
-}
-
-/* =========================== Class definitions =========================== */
-FixMLMOD::FixMLMOD(LAMMPS *lmp, int narg, char **arg) : Fix(lmp, narg, arg)
-{
-
-  const char *error_str_func = "FixMLMOD()";
-  
-  time_integrate = 1; /* set to 1 for fix performing integration, 0 if fix does not */
-  
-  /* setup the wrapper for MLMOD library */
-  wrapper_mlmod = new WrapperMLMOD(this,lmp,narg,arg);
-  
-}
-
-/* destructor */
-FixMLMOD::~FixMLMOD() {
-
-  /* write some final information to disk */
-  //writeFinalInfo();
-
-}
-
-void FixMLMOD::setup(int vflag) { /* lammps setup */
-
-  wrapper_mlmod->setup(vflag);
-  
-}
-
-/* ---------------------------------------------------------------------- */
[…]
