+++ src/src/USER-MLMOD/README.md
@@ -1,4 +1,4 @@
-For more information on mlmod package, see
+For more information on the mlmod package, see
 https://github.com/atzberg/mlmod and
 http://atzberger.org.
 
