+++ src/src/USER-MLMOD/examples/README.md
@@ -1,4 +1,4 @@
-For examples and more information, see
+For more information on the mlmod package, see
 https://github.com/atzberg/mlmod and
 http://atzberger.org.
 
