+++ examples/README.md
@@ -1,6 +1,6 @@
 __Examples__
 <p align="left">
-<img src="z_doc_img/particle1.png" width="20%"> 
+<img src="z_doc_img/particle1.gif" width="20%"> 
 </p>
 Please see the sub-folders above for example scripts.  Additional information and tutorials can be found at http://software.atzberger.org/
 
