+++ tests/mobility_responses1/template_response1/archive_gen_mlmod_rpy1.py
@@ -1,0 +1,199 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+# #### Creates a PyTorch model for testing the MLMOD-LAMMPS simulation codes
+# Paul J. Atzberger 
+# http://atzberger.org
+#     
+# Generates a hand-crafted model in PyTorch framework to show basic ideas. 
+#
+# This gives a stub for how to use the package and run simulation with 
+# them in LAMMPS
+#
+# These models can be replaced readily by data-driven / learned models or 
+# any other PyTorch model that can be traced and output to .pt torch-format.
+#
+import os, pickle;
+script_base_name = "gen_mlmod_rpy1";
+script_dir = os.getcwd();
+
+import torch
+import shutil, pickle;
+import numpy as np;
+import matplotlib;
+import matplotlib.pyplot as plt;
+import sys,shutil,pickle,pdb;
+
+import logging;
+
+def create_dir(dir_name):
+  if not os.path.exists(dir_name):
+    os.makedirs(dir_name);    
+
+# define the models (RPY)
+# RPY Tensor
+# """ M_ij = \frac{1}{6\pi\eta{a}}\left[
+# \frac{3a}{4r}\left(\mathcal{I} + \frac{rr^}{r}\right) + 
+# \frac{a^3}{2r^3}\left(\mathcal{I} - \frac{rr^}{r}\right)
+# \right]
+#
+# M_ii = \frac{1}{6\pi\eta{a}}\mathcal{I}
+#
+class M_ij_Model(torch.nn.Module):
+  def __init__(self):
+    super(M_ij_Model, self).__init__()
+    self.a = 5.1;
+    self.epsilon = (6.0/8.0)*self.a; self.eta = 1.0;
+    self.eps1 = 1e-12;
+
+  # evaluate
+  def forward(self, x):
+    num_dim = 3;
+
+    #epsilon = 6.0/8.0; eta = 1.0;  # can adjust below for coupling strength    
+    a = self.a; epsilon = self.epsilon; eta = self.eta; eps1 = self.eps1;
+    
+    vec_r_ij = x[0,0:num_dim] - x[0,num_dim:2*num_dim]; # avoiding tensorizing batch, just one pair assumed
+    r_ij_sq = torch.sum(torch.pow(vec_r_ij,2),0);
+    r_ij = torch.sqrt(r_ij_sq);
+    r_ij_p = r_ij + eps1;
+    r_ij_p_sq = r_ij_sq + 2*r_ij_p*eps1 + eps1*eps1; 
+    r_ij_p_cub = r_ij_p_sq*r_ij_p;
+
+    vec_r_ij_unit = vec_r_ij/r_ij_p;
+
+    c0 = 1.0/(6.0*torch.pi*eta*a);
+    c1 = 3.0*a/(4.0*r_ij_p);
+    M_ij = c1*(torch.eye(num_dim) + (torch.outer(vec_r_ij_unit,vec_r_ij_unit)));
+    c2 = a*a*a/(2.0*r_ij_p_cub)
+    M_ij = M_ij + c2*(torch.eye(num_dim) - 3.0*torch.outer(vec_r_ij_unit,vec_r_ij_unit));
+    M_ij = c0*M_ij;
+
+    z = M_ij.flatten().unsqueeze(0);
+
+    #pdb.set_trace();
+    return z;
+
+# define the model
+class M_ii_Model(torch.nn.Module):
+  def __init__(self):
+    super(M_ii_Model, self).__init__()
+    self.a = 5.1; self.eta = 1.0;
+        
+  # evaluate
+  def forward(self, x):
+    num_dim = 3;
+
[…]
