+++ tests/mobility_responses1/template_response1/archive_gen_mlmod_oseen1.py
@@ -1,0 +1,179 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+# #### Creates a PyTorch model for testing the MLMOD-LAMMPS simulation codes
+# Paul J. Atzberger 
+# http://atzberger.org
+#     
+# Generates a hand-crafted model in PyTorch framework to show basic ideas. 
+#
+# This gives a stub for how to use the package and run simulation with 
+# them in LAMMPS
+#
+# These models can be replaced readily by data-driven / learned models or 
+# any other PyTorch model that can be traced and output to .pt torch-format.
+#
+import os, pickle;
+script_base_name = "gen_mlmod_oseen1";
+script_dir = os.getcwd();
+
+import torch
+import shutil, pickle;
+import numpy as np;
+import matplotlib;
+import matplotlib.pyplot as plt;
+import sys,shutil,pickle,pdb;
+
+import logging;
+
+def create_dir(dir_name):
+  if not os.path.exists(dir_name):
+    os.makedirs(dir_name);    
+
+# define the model
+class M_ij_Model(torch.nn.Module):
+  def __init__(self):
+    super(M_ij_Model, self).__init__()
+    self.a = 5.1;
+    self.epsilon = (6.0/8.0)*self.a; self.eta = 1.0;
+    self.eps1 = 1e-12;
+
+  # evaluate
+  def forward(self, x):
+    num_dim = 3;
+    
+    vec_r_ij = x[0,0:num_dim] - x[0,num_dim:2*num_dim]; # avoiding tensorizing batch, just one pair assumed
+    r_ij_sq = torch.sum(torch.pow(vec_r_ij,2),0);
+    r_ij = torch.sqrt(r_ij_sq);
+
+    #epsilon = 6.0/8.0; eta = 1.0;  # can adjust below for coupling strength    
+    epsilon = self.epsilon; eta = self.eta; eps1 = self.eps1;
+    prefactor = 1.0/(8.0*torch.pi*eta*(r_ij + epsilon));
+    M_ij = prefactor*(torch.eye(num_dim) + (torch.outer(vec_r_ij,vec_r_ij)/(r_ij_sq + eps1)));
+
+    z = M_ij.flatten().unsqueeze(0);
+
+    #pdb.set_trace();
+    return z;
+
+# define the model
+class M_ii_Model(torch.nn.Module):
+  def __init__(self):
+    super(M_ii_Model, self).__init__()
+    self.a = 5.1; self.eta = 1.0;
+        
+  # evaluate
+  def forward(self, x):
+    num_dim = 3;
+
+    a = self.a; eta = self.eta;
+    M_ii = torch.eye(num_dim)*(1.0/(6.0*torch.pi*eta*a));
+                        
+    z = M_ii.flatten().unsqueeze(0);
+    
+    #pdb.set_trace();
+    return z;    
+
+# Initialize model
+M_ii_model = M_ii_Model();
+M_ij_model = M_ij_Model();
+
+a = 1.1; eta = 1.2; epsilon = (6.0/8.0)*a; eps1 = 1e-11; 
+M_ii_model.a = a; M_ii_model.eta = eta;
+M_ij_model.a = a; M_ij_model.eta = eta; 
+M_ii_model.epsilon = epsilon; M_ij_model.eps1 = eps1; 
+
+#-- save model data for later validation
+case_label = 'gen_001'
+base_dir = './output/%s/%s'%(script_base_name,case_label);
+
+print("base_dir = " + base_dir);
+create_dir(base_dir);
+
[…]
