+++ paper/paper.bib
@@ -483,6 +483,41 @@
   pages     = {83--94},
   volume    = {24},
   publisher = {Elsevier},
+  doi       = {10.1016/S1367-5788(00)90017-8}
+}
+
+
+@article{Plimpton:2022,
+title = {LAMMPS - a flexible simulation tool for particle-based materials
+modeling at the atomic, meso, and continuum scales},
+journal = {Computer Physics Communications},
+volume = {271},
+pages = {108171},
+year = {2022},
+issn = {0010-4655},
+doi = {10.1016/j.cpc.2021.108171},
+url = {https://www.sciencedirect.com/science/article/pii/S0010465521002836},
+author = {Aidan P. Thompson and H. Metin Aktulga and Richard Berger and Dan S.
+Bolintineanu and W. Michael Brown and Paul S. Crozier and Pieter J. {in 't
+Veld} and Axel Kohlmeyer and Stan G. Moore and Trung Dac Nguyen and Ray Shan
+and Mark J. Stevens and Julien Tranchida and Christian Trott and Steven J.
+Plimpton},
+keywords = {Molecular dynamics, Materials modeling, Parallel algorithms, LAMMPS},
+}
+
+@Article{Steigmann1999,
+  Title                    = {Fluid Films with Curvature Elasticity},
+  Author                   = {Steigmann, J. D.},
+  Journal                  = {Archive for Rational Mechanics and Analysis},
+  Year                     = {1999},
+  Number                   = {2},
+  Pages                    = {127--152},
+  Volume                   = {150},
+
+  Doi                      = {10.1007/s002050050183},
+  Owner                    = {atzberg},
+  Timestamp                = {2016.05.23},
+  Url                      = {http://dx.doi.org/10.1007/s002050050183}
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
