### 2023-08-29 #5620 @atzberg
Thanks for the feedback on the package.

I wanted to let you know I just pushed updates of the mlmod git repo with the latest version of things.  I'll write some more soon on the detailed items changed and items added in response to the comments above.

As a short summary, this includes a set of documentation pages, new scripts to simplify installation using the pre-compiled binaries, new scripts to help in compiling the package from source, and added codes for unit testing the package.   The documentation pages also include discussions of installation, compilation, information on the mlmod C++ and python APIs, and on the package usage.  The pre-compiled binaries are also now for lammps release Aug 2, 2023.  One challenge concerned builds handling the CXX11 ABI changes for backward compatibility with manylinux for pip installs.   I also made edits to the paper with the requested revisions.

I'll write some more details soon on the items.  I just wanted to mention updates are now pushed to the git repo.

Thanks again.
