+++ README.md
@@ -32,9 +32,9 @@ __To install directly__ the pre-compiled packages for Python, download one of th
 following
 
 - Linux Debian 9+/Ubuntu:
-  [mlmod_lammps-x.x.x-py3-none-manylinux_2_24_x86_64.whl](http://web.math.ucsb.edu/~atzberg/mlmod/distr/mlmod_lammps-0.8.0-py3-none-manylinux_2_24_x86_64.whl)
+  [mlmod_lammps-0.8.0-py3-none-manylinux_2_24_x86_64.whl](http://web.math.ucsb.edu/~atzberg/mlmod/distr/mlmod_lammps-0.8.0-py3-none-manylinux_2_24_x86_64.whl)
 - Linux Debian 9+/Ubuntu (flexible install):
-  [mlmod_lammps-x.x.x-py3-none-any.whl](http://web.math.ucsb.edu/~atzberg/mlmod/distr/mlmod_lammps-0.8.0-py3-none-any.whl)
+  [mlmod_lammps-0.8.0-py3-none-any.whl](http://web.math.ucsb.edu/~atzberg/mlmod/distr/mlmod_lammps-0.8.0-py3-none-any.whl)
 
 Install using 
 
