+++ requirements.txt
@@ -1,0 +1,2 @@
+numpy>=1.21.1
+
