+++ requirements.txt
@@ -1,4 +1,6 @@
 numpy>=1.21.1
-torch>=1.11.0
 mlmod-lammps @ http://web.math.ucsb.edu/~atzberg/mlmod/distr/mlmod_lammps-0.8.0-py3-none-any.whl
+
+# (optional) mlmod-lammps runs without pytorch (used for model generation)
+#torch>=1.11.0
 
