+++ examples/particles1/run_via_lmp1.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-# May need to edit the path and name for the mlmod-lammps executable "lmp".
-
-cd ./mlmod_model1
-lmp -in Model_lmp.LAMMPS_script
-
