+++ examples/particles1/vis_pv1.py
@@ -1,0 +1,232 @@
+# state file generated using paraview version 5.9.1
+
+#### import the simple module from the paraview
+from paraview.simple import *
+#### disable automatic camera reset on 'Show'
+paraview.simple._DisableFirstRenderCameraReset()
+
+import glob
+
+#output_dir = './output/run_proto_test3/batch_00_test001/vtk';
+output_dir = './output/run_sim_particles1/batch_00_test001/vtk';
+print("output_dir = " + output_dir);
+
+# ----------------------------------------------------------------
+# setup views used in the visualization
+# ----------------------------------------------------------------
+
+# get the material library
+materialLibrary1 = GetMaterialLibrary()
+
+# Create a new 'Render View'
+flag_view='view1';
+if flag_view=='view1':
+  # Create a new 'Render View'
+  renderView1 = CreateView('RenderView')
+  renderView1.ViewSize = [1241, 479]
+  renderView1.AxesGrid = 'GridAxes3DActor'
+  renderView1.StereoType = 'Crystal Eyes'
+  renderView1.CameraFocalDisk = 1.0
+  renderView1.BackEnd = 'OSPRay raycaster'
+  renderView1.OSPRayMaterialLibrary = materialLibrary1
+
+elif flag_view=='view2':
+  renderView1 = CreateView('RenderView')
+  renderView1.ViewSize = [1166, 640]
+  renderView1.AxesGrid = 'GridAxes3DActor'
+  renderView1.StereoType = 'Crystal Eyes'
+  renderView1.CameraPosition = [0.0, 0.0, 67.99559695559125]
+  renderView1.CameraFocalDisk = 1.0
+  renderView1.CameraParallelScale = 31.176914536239792
+  renderView1.BackEnd = 'OSPRay raycaster'
+  renderView1.OSPRayMaterialLibrary = materialLibrary1
+
+
+SetActiveView(None)
+
+# ----------------------------------------------------------------
+# setup view layouts
+# ----------------------------------------------------------------
+
+# create new layout object 'Layout #1'
+layout1 = CreateLayout(name='Layout #1')
+layout1.AssignView(0, renderView1)
+layout1.SetSize(1166, 640)
+
+# ----------------------------------------------------------------
+# restore active view
+SetActiveView(renderView1)
+# ----------------------------------------------------------------
+
+# ----------------------------------------------------------------
+# setup the data processing pipelines
+# ----------------------------------------------------------------
+
+# create a new 'XML PolyData Reader'
+particles_file_list = glob.glob(output_dir + '/' + 'Particles*.vtp');
+particles_file_list.sort();
+num_particles_file_list = len(particles_file_list);
+print("num_particles_file_list = " + str(num_particles_file_list));
+
+particlesData = XMLPolyDataReader(registrationName='particles', FileName=particles_file_list)
+particlesData.PointArrayStatus = ['id', 'type', 'vx', 'fx']
+particlesData.TimeArray = 'None'
+
+# create a new 'XML Unstructured Grid Reader'
+bounding_box_file_list = glob.glob(output_dir + '/' + 'Particles*boundingBox.vtu');
+bounding_box_file_list.sort();
+num_bounding_box_file_list = len(bounding_box_file_list);
[…]
