+++ examples/particles1/vis_pv1.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+paraview --script=vis_pv1.py
+
+

[…]
