+++ examples/README.md
@@ -2,5 +2,7 @@ __Examples__
 <p align="left">
 <img src="z_doc_img/particle1.png" width="20%"> 
 </p>
-Please see the sub-folders above for examples setting up models and performing simulations.  Additional information can also be found in the [documentation pages](http://web.math.ucsb.edu/~atzberg/mlmod/docs/index.html).
+
+Please see the sub-folders above for examples setting up models and performing simulations.  Additional information can also be found in the
+[documentation pages](http://web.math.ucsb.edu/~atzberg/mlmod/docs/index.html).
 
