+++ tests/README.md
@@ -1,0 +1,2 @@
+Please see the sub-folders above for running tests.  Additional information can also be found in the
+[documentation pages](http://web.math.ucsb.edu/~atzberg/mlmod/docs/index.html).
