+++ images/README.md
@@ -1,0 +1,1 @@
+Folder for images used in the documents.
