### 2023-08-30 #5620 @atzberg
@JColdstream (detailed responses to comments below)

* "On Ubuntu 22.04: The installation worked nicely with the python package.

  Installing it from source was also OK, the compilation instructions in ./src/INSTALL.md are good. There are also some (old?)
  instructions in ./lib/mlmod/INSTALL.md which could be updated to be the same as ./src/INSTALL.md."

We have now created centralized documentation pages for installing the package and for compiling from the source codes.  We also had to grapple with some issues with CXX11 ABI when updating for the latest release of lammps (Aug 2, 2023).  We have put together both helper scripts which allow for automating much of the build process.  We also put together and re-organized the Makefiles clarifying how to specify the library dependencies, paths, and other settings.  For clarity and robustness, this also now includes special flags for abi and other settings.  The new makfiles can be found in the $mlmod_git/src directories, and discussed on the documentation pages linked in directory $mlmod_git/doc.

We also now give in the document pages discussion of two approaches to building the package (i) using the automated scripts, and (ii) performing each of the build steps manually.   Again, for the typical user, much of the compilation process is not needed, since they can use the pre-compiled binaries and 'pip install -U (mlmod-pkg).whl".

As for PyPI and pre-compiled binaries, there is a 100MB limit, so we were unable to host our wheels (.whl files) there currently, and we instead host the .whl on our own website.
We now give detailed instructions on the github page and document pages for this installation.  To further help, we also put together a quick_install.py script which further automates the installation process.  The script can be found in the home directory $mlmod_git/.

* "The only issue was I had to edit the environment variable LD_LIBRARY_PATH to get it to run, but could just be a product of
  where I installed the dependencies.

  "./lmp_mlmod_serial: error while loading shared libraries: libmlmod.so: cannot open shared object file: No such file or
  directory"

  and

  "./lmp_mlmod_serial: error while loading shared libraries: libtorch_cpu.so: cannot open shared object file: No such file or
 directory"

  Was fixed by pointing to them at runtime with:
  export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/path/to/mlmod/:/path/to/libtorch/lib/"

To help mitigate these types of issues with shared library locations, versions, and other issues, we used in our pre-compiled binaries an alternative approach based on r-paths and local copies of the needed libraries.  When building the package manually, we have also now re-organized the makefiles and path variables to make more explicit the dependencies and locations of the shared libraries.  Fortunately, for distributions such as Ubuntu, the needed libraries get installed in rather standard locations already in the search path.

For mlmod-lammps, we have now added comments to the documentation pages on the paths you mention above and discuss how to adjust the LD_LIBRARY_PATH.  The new document pages can be found in the links in the directory $mlmod_git/doc.

* "The example runs and gives me a nice trajectory, but have you got any checks to make sure the code is actually giving the right answer?  Either something similar to pytest or something like an example calculation/dataset to compare to maybe? Just looking for something for the Automated tests section in the review."

Thanks for this feedback.  We have now put together pytests for the package.  For example for configurations of particles we look at the mlmod computed mobility responses and make comparisons with analytic mobility responses.  This can be found in the directory $mlmod_git/tests.

* "Other things: I really like the little intro video on YouTube."

Thanks, we have now included a link to this overview in the documentation pages.

* "LAMMPS has a new citation that they prefer to be used (https://doi.org/10.1016/j.cpc.2021.108171) rather than the old Plimpton 1995 one."

Thanks for bringing this to my attention.  We have now revised our paper with the updated citation and DOI for the new lammps paper.

In addition to the items above, we have also put together detailed documentation pages.  This includes information on package installation, compiling the package and configuration, and on example usage.   The documentation also discusses some of the C++ API and related python interface to the package via lammps.  We also revised the examples, which now simplify the process to generate ml models and the process in setting up simulations.

Thanks again for this feedback on the package.
