### 2023-08-30 #5620 @atzberg
@RuiApostolo (detailed responses to comments below)

> I'll try to organise my recommendations in relevant sections
>
> # Installation
> ## pip
> The installation using pip is really simple, but it contains an older version of LAMMPS (29 Oct 2020). If possible, it would be preferable to upload the wheel file to pypi.org so that users could simply `pip install mlmod`.
>

We have now packaged pre-compiled binaries for the latest release of lammps (Aug 2, 2023).  This presented some challenges to build given the CXX11 ABI and aim to be backward compatible with manylinux distributions.  We have also simplified the installation process.  This includes an automated script quick_install.py available on the homepage $mlmod_git/ and step-by-step instructions on the homepage and documentation pages.  As for PyPI, we are running into the 100MB for direct distribution there, so we are hosting the pip wheels .whl on our website.   Our new simplified installation process now only requires a single command like: `pip (package-url-name)`.

> # Compiling
> Compiling MLMOD from source has quite a few steps, and while it's not overly complicated, I would not recommend it for anyone not used to compiling software with (mildly) complex dependencies. It could be made easier if there was a more detailed example script containing all the commands and some example paths. In Ubuntu 22.04, FFTW3 (libfftw3-dev), LIBLAPACKE (liblapacke-dev), and tinyxml2 (libtinyxml2-dev) are available to install as packages, leaving the user to install only libtorch (which is very simple) and the MLMOD-patched version of LAMMPS.

We have now re-organized the build process simplifying the process.  We also put together scripts to help automate many of the build steps.  This includes copying the files to needed locations in lammps and calling the needed makefiles for the library and overall build.  We also put together detailed documentation pages discussing including two approaches to building the package (i) using our automatic scripts, and (ii) manually performing the steps of the build.   Our re-organized makefiles and documentation pages also clarify how the build relates to dependent libraries and how these can be adjusted.   We also added flags for more robustly handling different abis and system settings.   The build process for compiling from source is discussed in the documentation pages in directory $mlmod_git/doc

>
> With the recent change in LAMMPS development choice to provide a [newbie-friendly `cmake` compilation instructions rather than make](https://docs.lammps.org/Build_cmake.html#advantages-of-using-cmake), it would be useful to provide instructions for compiling MLMOD using `cmake`. Furthermore, having a shell script that copies all the files MLMOD needs (src/USER-MLMOD, lib, MAKE) to a user-specified location where the LAMMPS source files are would remove several steps and sources of potential error.

Thanks for this suggestion to further abstract and automate the build process with cmake.   There are a number of subtle dependencies in lammps and configuration scripts we currently utilize for the package during builds.  We will look into abstracting this further to utilize the cmake paradigm to generate the Makefiles and run the build process in future releases.

>Finally, consider submitting the USER-MLMOD package to the [LAMMPS repository](https://github.com/lammps/lammps/blob/6422565048106045443f022b26bd144d23505533/.github/CONTRIBUTING.md), so that it may be included in the general release of LAMMPS.

This is also something we plan to look into.   Thanks for mentioning.

>
> ## README
> On the main README, you should link the sentence "For more information on other ways to install or compile the package, please see the documentation page." to the `src/INSTALL.md` file.
>

We have now put together centralized documentation pages.  This information is now located there with details on the installation process, both using the pre-compiled binaries and for building the package from source codes.   We also put together scripts for helping to automate some steps of installation and the build process.

> # Tests
> I agree with @JColdstream here, some tests to make sure that the package is returning expected results would be very useful. It could help identify compatibility issues with certain LAMMPS or library versions, should they exist.
>

We have now put together pytests for the codes.   For example for configurations of particles we look at the mlmod computed mobility responses and make comparisons with analytic mobility responses. This can be found in the directory $mlmod_git/tests.

> # Software documentation
> I think this needs to be created. At a minimum, there needs to be documentation on the LAMMPS commands that MLMOD introduces, and all possible parameters they accept. Ideally, these would follow the same structure that the LAMMPS documentation already has, so that the familiar structure would help users already familiar with LAMMPS.

We have now put together documentation pages in style similar to lammps.  We now discuss also usage of the mlmod package, which is cast in the style of a "lammps fix."  We also give some discussion of the C++ APIs and the python interfaces to the mlmod-lammps package.   These can be found in the directory $mlmod_git/doc

> # Paper
> Overall well written, I have only found a few minor typos (I'm providing the lines as numbered on the PDF created by the bot above):
>
> * `accomodated` line 32
> * `accomodate` line 59
> * `simiplified` line 93
> * `Nielsen:2000` line 110 (missing `@` to make it a reference, I believe)

Thanks.  We have now revised the paper taking the comments above into account.  We have also resolved the DOI related references.

In addition to the items above, we have also revised the examples, which now simplify the process to generate ml models and the process in setting up simulations.  We also put together scripts for automating many of the steps for installation and for building the package.

Thanks again for this feedback on the package.
