### 2023-08-30 #5620 @RuiApostolo
@atzberg Incredible work on the changes, I think the new scripts for installing and compiling will make things much easier for anyone wanting to use MLMOD.
I didn't know about the 100MB limit on pypi, your solution makes sense.

@ppxasjsm From my end, this looks good, I recommend acceptance.
