### 2023-08-30 #5620 @JColdstream
@atzberg Good stuff! Compilation is much smoother now and the tests are simple and easy to use. The new documentation also looks good.

@ppxasjsm Looks good to go.
