+++ paper/paper.bib
@@ -424,14 +424,14 @@
 }
 
 @Article{zenodo,
-  author       = {Paul J. Atzberger},
+  author       = {P. J. Atzberger},
   journal      = {Zenodo},
-  title        = {MLMOD Package v1.0.0},
+  title        = {MLMOD Package v1.0.1},
   year         = {2023},
   abstractnote = {MLMOD Package},
-  doi          = {10.5281/zenodo.7897824},
+  doi          = {10.5281/zenodo.8327516},
   publisher    = {Zenodo},
-  url          = {http://dx.doi.org/10.5281/zenodo.7897824},
+  url          = {https://doi.org/10.5281/zenodo.8327516}
 }
 
 @inproceedings{Kingma:2014,
