+++ README.md
@@ -3,7 +3,7 @@
 </p>
 
 [Documentation](http://web.math.ucsb.edu/~atzberg/mlmod/docs/index.html) |
-[Examples](https://github.com/atzberg/mlmod/tree/master/examples) |
+[Examples](./examples) |
 [Paper](https://arxiv.org/abs/2107.14362)
                                                                                                 
 ### MLMOD: Machine Learning Methods for Data-Driven Modeling in LAMMPS
@@ -33,9 +33,9 @@ __To install directly__ the pre-compiled packages for Python, download one of th
 following
 
 - Linux Debian 9+/Ubuntu:
-  [mlmod_lammps-0.8.0-py3-none-manylinux_2_24_x86_64.whl](http://web.math.ucsb.edu/~atzberg/mlmod/distr/mlmod_lammps-0.8.0-py3-none-manylinux_2_24_x86_64.whl)
+  [mlmod_lammps-0.8.0-py3-none-manylinux_2_24_x86_64.whl](http://web.math.ucsb.edu/~atzberg/mlmod/distr/mlmod_lammps-1.0.1-py3-none-manylinux_2_24_x86_64.whl)
 - Linux Debian 9+/Ubuntu (flexible install):
-  [mlmod_lammps-0.8.0-py3-none-any.whl](http://web.math.ucsb.edu/~atzberg/mlmod/distr/mlmod_lammps-0.8.0-py3-none-any.whl)
+  [mlmod_lammps-0.8.0-py3-none-any.whl](http://web.math.ucsb.edu/~atzberg/mlmod/distr/mlmod_lammps-1.0.1-py3-none-any.whl)
 
 Install using 
 
@@ -47,7 +47,7 @@ To test the package installed run
 
 __For example models and simulations__, see the notebooks and scripts in the
 [examples
-folder](https://github.com/atzberg/mlmod/tree/master/examples).  
+folder](/examples).  
 
 __Documentation pages__ can be found
 [here](http://web.math.ucsb.edu/~atzberg/mlmod/docs/index.html).
@@ -133,7 +133,7 @@ http://atzberger.org/
 
 ----
 [Documentation](http://web.math.ucsb.edu/~atzberg/mlmod/docs/index.html) |
-[Examples](https://github.com/atzberg/mlmod/tree/master/examples) |
+[Examples](./examples) |
 [Atzberger Homepage](http://atzberger.org/)
 
 
