### 2023-09-08 #5620 @atzberg
> Hi @atzberg can you please complete this checklist:

> * [x]  Double check authors and affiliations (including ORCIDs)

> * [x]   Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

Release for package version 1.0.1 has been posted as tag v1.0.1

> * [x]   Archive the release on Zenodo/figshare/etc and post the DOI here.

DOI:
10.5281/zenodo.8327516

URL:
https://doi.org/10.5281/zenodo.8327516

> * [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

> * [x]  Make sure that the license listed for the archive is the same as the software license.

I have updated the GitHub with these items.   The checklist has now been completed.

Just let me know if you need anything else.
