+++ examples/force1/run_sim_force1.py
@@ -1,0 +1,347 @@
+#!/usr/bin/env 
+### Particle simulation with mlmod-based mobility.
+
+import os,sys, shutil, pickle, ipdb;
+import numpy as np;
+
+script_base_name = "run_sim_force1"; script_dir = os.getcwd();
+
+# import the mlmod_lammps module
+from mlmod_lammps.lammps import lammps # use this for the pip install of pre-built package
+lammps_import_comment = "from mlmod_lammps.lammps import lammps";  
+
+# filesystem management
+def create_dir(dir_name):
+  if not os.path.exists(dir_name):
+    os.makedirs(dir_name);    
+    
+def rm_dir(dir_name):
+  if os.path.exists(dir_name):    
+    shutil.rmtree(dir_name);
+  else: 
+    print("WARNING: rm_dir(): The directory does not exist, dir_name = " + dir_name);    
+
+def copytree2(src, dst, symlinks=False, ignore=None):
+  for ff in os.listdir(src):
+    s = os.path.join(src, ff); d = os.path.join(dst, ff);
+    if os.path.isdir(s):
+      shutil.copytree(s, d, symlinks, ignore);
+    else:
+      shutil.copy2(s, d);
+
+def write_mlmod_F_ML1_params(filename,params):
+  model_type = params['model_type']; model_data = params['model_data'];
+  base_name, base_dir, mask_fix, mask_input, F_filename = tuple(map(model_data.get,\
+                                  ['base_name', 'base_dir', 'mask_fix', 
+                                   'mask_input', 'F_filename']));
+  
+  # xml file
+  f = open(filename,'w');
+  f.write('<?xml version="1.0" encoding="UTF-8"?>\n');
+  f.write('<MLMOD>\n');
+  f.write('<model_data type="' + model_type + '">\n');
+  f.write('<base_name value="' + base_name + '"/>\n');
+  f.write('<base_dir value="' + base_dir + '"/>\n');
+  f.write('<mask_fix value="' + mask_fix + '"/>\n');
+  f.write('<mask_input value="' + mask_input + '"/>\n');
+  f.write('<F_filename value="' + F_filename + '"/>\n');
+  f.write('</model_data>\n');
+  f.write('</MLMOD>\n');
+  f.close();
+
+  # pickle file
+  f = open(filename + '.pickle','wb'); pickle.dump(params,f); f.close();
+
+def write_mlmod_Dyn_ML1_params(filename,params):
+  model_type = params['model_type']; model_data = params['model_data'];
+  base_name, base_dir, mask_fix, mask_input, dyn1_filename, dyn2_filename = tuple(map(model_data.get,\
+                                  ['base_name', 'base_dir', 'mask_fix', 
+                                   'mask_input', 'dyn1_filename', 'dyn2_filename']));
+  
+  # xml file
+  f = open(filename,'w');
+  f.write('<?xml version="1.0" encoding="UTF-8"?>\n');
+  f.write('<MLMOD>\n');
+  f.write('<model_data type="' + model_type + '">\n');
+  f.write('<base_name value="' + base_name + '"/>\n');
+  f.write('<base_dir value="' + base_dir + '"/>\n');
+  f.write('<mask_fix value="' + mask_fix + '"/>\n');
+  f.write('<mask_input value="' + mask_input + '"/>\n');
[…]
