+++ examples/dynamics1/README.md
@@ -2,7 +2,7 @@
 <img src="doc_img/dynamics1.png" width="15%"> 
 </p>
 
-Simulation using dynamics from a pytorch model with X^{n+1} = A(X^n,V^n,F^n).
+Simulation using dynamics from a pytorch model $A(\cdot,\theta)$ with $X^{n+1} = A(X^n,V^n,F^n;\theta)$.
 
 To run the simulation for the python mlmod package, use 
 
