### 2023-09-12 aa9c99f
setup for testing different force types
