### 2023-09-12 #5620 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
