### 2023-09-14 ac55a5c
adding citation information
