+++ README.md
@@ -4,7 +4,7 @@
 
 [Documentation](http://web.math.ucsb.edu/~atzberg/mlmod/docs/index.html) |
 [Examples](./examples) |
-[Paper](https://arxiv.org/abs/2107.14362)
+[Paper](https://joss.theoj.org/papers/cad4ad083ae0cef287faca542db5229d)
                                                                                                 
 ### MLMOD: Machine Learning Methods for Data-Driven Modeling in LAMMPS
 Now available with Jupyter notebooks and Python scripts for readily setting up
@@ -82,17 +82,21 @@ pages](http://web.math.ucsb.edu/~atzberg/mlmod/docs/index.html).
 Please cite the paper below for this package:
 
 **MLMOD Package: Machine Learning Methods for Data-Driven Modeling in
-LAMMPS"**, P.J. Atzberger, arXiv.2107.14362, 2021 [[paper
-link]](https://arxiv.org/abs/2107.14362).
+LAMMPS"**, P.J. Atzberger, Journal of Open Source Software, 8(89), 5620, (2023) [[paper
+link]](https://doi.org/10.21105/joss.05620).
 ```
 @article{mlmod_atzberger,
-  author  = {Atzberger, P. J.},
-  journal = {arxiv},
-  title   = {MLMOD Package: Machine Learning Methods for Data-Driven Modeling in LAMMPS},
-  year    = {2021},
-  note    = {http://atzberger.org},
-  doi     = {10.48550/arXiv.2107.14362},
-  url     = {https://arxiv.org/abs/2107.14362},
+  author    = {Paul J. Atzberger},
+  journal   = {Journal of Open Source Software}, 
+  title     = {MLMOD: Machine Learning Methods for Data-Driven Modeling in LAMMPS},
+  year      = {2023},  
+  publisher = {The Open Journal},
+  volume    = {8},
+  number    = {89},
+  pages     = {5620},
+  note      = {http://atzberger.org},
+  doi       = {10.21105/joss.05620},
+  url       = {https://doi.org/10.21105/joss.05620}
 }
 ```
 ----
