+++ examples/qoi1/gen_and_setup_qoi1.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+python ./gen_mlmod_qoi1.py
+
+./cp_qoi1.sh
+
+
