+++ examples/qoi1/cp_qoi1.sh
@@ -1,0 +1,9 @@
+#!/bin/bash -x
+
+dir1=./output/gen_mlmod_qoi1/gen_001
+dir2=./mlmod_model1
+
+mkdir -p $dir2
+cp $dir1/*QoI_qoi1* $dir2
+
+
