+++ examples/particles1/gen_and_setup_particles1.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+python ./gen_mlmod_oseen1.py
+python ./gen_mlmod_rpy1.py
+
+./cp_particles1.sh
+
+
