+++ examples/force1/README.md
@@ -1,3 +1,14 @@
-.
+<p align="left">
+<img src="doc_img/force1.png" width="15%"> 
+</p>
 
+Simulation using a force that acts on particles from a pytorch model $F(X)$.
+
+To run the simulation for the python mlmod package, use 
+
+```python run_sim_force1.py```
+
+To generate the PyTorch models, use 
+
+```python gen_mlmod_force1.py```
 
