+++ examples/dynamics1/README.md
@@ -1,3 +1,14 @@
-.
+<p align="left">
+<img src="doc_img/dynamics1.png" width="15%"> 
+</p>
 
+Simulation using dynamics from a pytorch model $A(\cdot;\theta)$ with $X^{n+1} = A(X^n,V^n,F^n;\theta)$.
+
+To run the simulation for the python mlmod package, use 
+
+```python run_sim_dynamics1.py```
+
+To generate the PyTorch models, use 
+
+```python gen_mlmod_dynamics1.py```
 
