+++ src/python/mlmod_lammps/util.py
@@ -1,0 +1,518 @@
+r"""Utility routines.  Paul J. Atzberger, http://atzberger.org/ """;
+
+import os,sys,shutil,pickle,logging;
+import numpy as np;
+
+# setup default print_log statement
+global print_log;
+print_log = print;
+
+# lammps related
+def print_version_info(L,print_log=print):
+  #L.command("info all out overwrite tmp.txt"); # dump all current info to file.
+  L.command("info all out overwrite tmp.txt"); # dump all current info to file.
+  f = open("tmp.txt"); ss = f.read(); f.close(); # load a display for the user
+  results = [];
+  for line in ss.splitlines():
+    if "version" in line:
+      print_log(line);   
+
+# filesystem management
+def create_dir(dir_name):
+  if not os.path.exists(dir_name):
+    os.makedirs(dir_name);    
+    
+def rm_dir(dir_name):
+  if os.path.exists(dir_name):    
+    shutil.rmtree(dir_name);
+  else: 
+    print_log("WARNING: rm_dir(): The directory does not exist, dir_name = " + dir_name);    
+
+def copytree2(src, dst, symlinks=False, ignore=None):
+  for ff in os.listdir(src):
+    s = os.path.join(src, ff); d = os.path.join(dst, ff);
+    if os.path.isdir(s):
+      shutil.copytree(s, d, symlinks, ignore);
+    else:
+      shutil.copy2(s, d);
+    
+# plotting
+def save_fig(base_filename,extraLabel='',flag_verbose=True,dpi_set=200,flag_pdf=False,bbox_inches=None,print_log=print):
+    
+  import matplotlib;
+  import matplotlib.pyplot as plt;
+
+  flag_simple = False;
+  if flag_simple: # avoids flashing, just uses screen resolution (ignores dpi_set)
+    if flag_pdf:    
+      save_filename = '%s%s.pdf'%(base_filename,extraLabel);
+      if flag_verbose:
+        print_log('save_filename = %s'%save_filename);
+      #plt.savefig(save_filename, format='pdf',dpi=dpi_set,facecolor=(1,1,1,1),alpha=1.0);
+      plt.savefig(save_filename, format='pdf',bbox_inches=bbox_inches);
+
+    save_filename = '%s%s.png'%(base_filename,extraLabel);
+    if flag_verbose:
+      print_log('save_filename = %s'%save_filename);
+    #plt.savefig(save_filename, format='png',dpi=dpi_set,facecolor=(1,1,1,1),alpha=1.0);
+    plt.savefig(save_filename, format='png',bbox_inches=bbox_inches);
+  else:  # NOTE: Can also get same below likely by setting facecolor='white'.
+    fig = plt.gcf();  
+    fig.patch.set_alpha(1.0);
+    fig.patch.set_facecolor((1.0,1.0,1.0,1.0));
+    
+    if flag_pdf:
+      save_filename = '%s%s.pdf'%(base_filename,extraLabel);
+      if flag_verbose:
+        print_log('save_filename = %s'%save_filename);
+      plt.savefig(save_filename, format='pdf',dpi=dpi_set,facecolor=(1,1,1,1),bbox_inches=bbox_inches);
+      #plt.savefig(save_filename, format='pdf',dpi=dpi_set);
+
+    save_filename = '%s%s.png'%(base_filename,extraLabel);
+    if flag_verbose:
+      print_log('save_filename = %s'%save_filename);
+    plt.savefig(save_filename, format='png',dpi=dpi_set,facecolor=(1,1,1,1),bbox_inches=bbox_inches);
+    #plt.savefig(save_filename, format='png',dpi=dpi_set);
+
+    
+# logging 
[…]
