+++ src/python/mlmod_lammps/examples/test1.py
@@ -1,0 +1,9 @@
+# quick test of package
+# see http://atzberger.org/ for examples and more information
+
+from selm_lammps.lammps import PyLammps
+
+L = PyLammps();
+
+L.command("info all");
+
