+++ src/python/mlmod_lammps/lammps/lib/README.md
@@ -1,0 +1,1 @@
+Placeholder to put additional shared libraries and dependencies. 
