+++ src/python/mlmod_lammps/lammps/README.md
@@ -1,0 +1,1 @@
+Placeholder to put files for lammps package, python interface, and shared libraries. 
