+++ examples/qoi1/README.md
@@ -1,3 +1,14 @@
-.
+<p align="left">
+<img src="doc_img/qoi1.png" width="15%"> 
+</p>
 
+Simulation using machine learning methods to compute quantities of interest, $A(X,V,F,I_T,t)$.
+
+To run the simulation for the python mlmod package, use 
+
+```python run_sim_qoi1.py```
+
+To generate the PyTorch models, use 
+
+```python gen_mlmod_qoi1.py```
 
