+++ examples/mpi1/gen_mlmod_force1.py
@@ -1,0 +1,645 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+# #### Creates a PyTorch model for testing the MLMOD-LAMMPS simulation codes
+# Paul J. Atzberger 
+# http://atzberger.org
+#     
+# Generates model in PyTorch framework to show basic ideas. 
+#
+# These models can be replaced readily by almost any PyTorch model that 
+# can be traced and output to .pt torch-format (neural networks, gpr, etc...).
+import os, pickle;
+script_base_name = "gen_mlmod_force1";
+script_dir = os.getcwd();
+
+import torch
+import shutil, pickle;
+import numpy as np;
+import matplotlib;
+import matplotlib.pyplot as plt;
+import sys,shutil,pickle,pdb;
+
+import logging;
+
+def create_dir(dir_name):
+  if not os.path.exists(dir_name):
+    os.makedirs(dir_name);    
+
+# define force models
+#
+class F_ML1_Model(torch.nn.Module):
+
+  def __init__(self,**params):
+    super(F_ML1_Model, self).__init__()
+
+    # tries to get or sets variable to None
+    self.a, self.num_dim = tuple(map(params.get,
+                           ['a','num_dim']));
+
+    if self.num_dim is None:
+      self.num_dim = 3;
+
+    self.mask_input,self.num_atoms = tuple(map(params.get,
+                           ['mask_input','num_atoms']));
+
+    self.list_mask = list_mask = self.mask_input.split();
+    self.num_inputs = num_inputs = len(list_mask);
+
+    self.max_num_inputs = max_num_inputs = 5;
+    mask_input_flag = [False]*max_num_inputs;
+       
+    # detect which flags are set (organized in order) 
+    for i in range(0,self.num_inputs):
+      if list_mask[i] == "X":
+        mask_input_flag[0] = True;
+      elif list_mask[i] == "V":
+        mask_input_flag[1] = True;
+      elif list_mask[i] == "F":
+        mask_input_flag[2] = True;
+      elif list_mask[i] == "Type":
+        mask_input_flag[3] = True;
+      elif list_mask[i] == "Time":
+        mask_input_flag[4] = True;
+      else:
+        print("No recognized, list_mask[%d] = %s"%(i,self.list_mask));
+    
+    self.mask_input_flag = mask_input_flag; 
+
+    # get dimensions of each segment of the data
+    # and set index ranges  
+    i_dim = []; num_dim = self.num_dim; I0 = 0;     
+    if mask_input_flag[0]:  # X
+      i_dim.append(num_dim);
+    else:
+      i_dim.append(0);
+
+    if mask_input_flag[1]:  # V
+      i_dim.append(num_dim);
+    else:
+      i_dim.append(0);
+
+    if mask_input_flag[2]:  # F
+      i_dim.append(num_dim);
+    else:
+      i_dim.append(0);
+
+    if mask_input_flag[3]:  # Type
+      i_dim.append(int(1));
+    else:
+      i_dim.append(0);
+
+    if mask_input_flag[4]:  # time
+      i_dim.append(int(1));
+    else:
+      i_dim.append(0);
+
+    II1 = []; II2 = []; I = 0; # setup ranges
+    for k in np.arange(0,self.max_num_inputs - 1): # X V F Type
+      d = i_dim[k];
+      II1.append(I); I += num_atoms*d; II2.append(I);
+
[…]
