+++ examples/mpi1/mpi_force1.py
@@ -1,12 +1,13 @@
-#!/home/atzberg/anaconda3/envs/mlmod-lammps/bin/python -i
-
-# preceding line should have path for Python on your machine
-
-# base on mpi simple.py
-
-# Parallel syntax: mpirun -np 4 code.py 
+#!/bin/python
+#
+# parallel simulations use:
+#
+#   mpirun -np 4 python code.py 
+#
+# based on mpi lammps simple.py
 #
 # uses package: mpi4py 
+#
 
 from __future__ import print_function
 import os,sys,pickle;
@@ -21,25 +22,17 @@ def wrap_print1(fid,ss):
   fid.write(ss + '\n');
   print(ss);
 
-# parse command line
-
-#argv = sys.argv
-#if len(argv) != 2:
-#  print("Usage: python %s.py in.lammps"%script_base_name)
-#  sys.exit()
-
-#infile = sys.argv[1]
 infile='in.force1';
 
 me = 0
 
-# uncomment this if running in parallel via mpi4py
+# mpi4py
 from mpi4py import MPI
+
 me = MPI.COMM_WORLD.Get_rank()
 nprocs = MPI.COMM_WORLD.Get_size()
 
-#from lammps import lammps
-# import the mlmod_lammps module
+# import mlmod-lammps
 from mlmod_lammps.lammps import lammps; 
 import mlmod_lammps.lammps.constants as lconst
 import mlmod_lammps.util as m_util;
@@ -52,8 +45,6 @@ if flag_wrap_print:
   printw("proc_filename = " + proc_filename);
 else:
   printw = lambda ss: wrap_print0(ss);
-
-# wait for me == 0 to finish setup before other processes proceed
 
 force_case = 'force1';
 base_dir_output = '%s/output/%s'%(script_dir,script_base_name);
@@ -96,9 +87,7 @@ printw("."*80);
 
 # setup mlmod force 
 flag_force=True;
-#force_type='F_ML1';
 force_type='F_X_ML1';
-#force_type='F_Pair_ML1';
 print("force_type = " + force_type);
 if flag_force: 
   f_basename = 'F_' + force_case + '_' + force_type;
@@ -116,7 +105,7 @@ if flag_force:
   filename_mlmod_params = force_type + '_' + force_case + '.mlmod_params';
   if me == 0:  # just write file for rank == 0 process (assumes shared filesystem)
     m_util.write_mlmod_params(filename_mlmod_params,mlmod_params);
-  #lmp.command("fix F_ML1_" + force_case " all mlmod " + filename_mlmod_params);
+
   lmp.command("fix %s_1 all mlmod"%force_type + " " + filename_mlmod_params + " verbose 2");
 
 lmp.command("variable dumpfreq equal 1");
@@ -125,15 +114,9 @@ lmp.command("dump_modify dvtk_mlmod1 pad 8"); # ensures filenames file_000000.da
 lmp.command("dump_modify dvtk_mlmod1 sort id");
 
 printw("."*80);
-# run 10 more steps
-# get coords from LAMMPS
-# change coords of 1st atom
-# put coords back into LAMMPS
-# run a single step with changed coords
 
 lmp.command("run 300")
 
-# uncomment if running in parallel via mpi4py
 printw("Proc %d out of %d procs has"%(me,nprocs))
 printw("Proc %d: Done."%me);
 
